Used /dev/urandom as Seed Rather Than Current Time.
I was running into trouble when I ran the command more than once in a second.
I would invariably get the same result both times.

Most of the time this wasn't a problem, but when I wanted to populate a playlist with 5 songs, it'd be best if they weren't all the same.

So, now I try to use /dev/urandom.

In my testing, I did find it was slightly less random, strangely.
When I ran 100 entries of 1-5 the distribution was slanted towards 3.
I'm not sure, mathemtically why that would be.
I am, after all, using modulus. This means that the number that come out are on average more likely to be 3 more than a multiple of 5.
That seems kind of weird and arbitrary.
The only thing I can think of is that there's a certain value that comes up more often, and it is 3 more.

Either way, this is how it is.
When using time, the distribution seemed fair.
You could only test it by putting a sleep 1 into every iteration, though, because otherwise all 100 would be the same value.

So... trade off...