Pick Can Now Pick More Than 1 Item.
Running Pick over the same input multiple times to pick multiple values was dumb.
There were problems with randomness, and it was inefficient.

Now, the recommended way is to run 'pick n' instead.

Probably one of the more confusing changes in this patch is the difference in what the variable "current" does.
Previously "current" held the current value that was to be output, and "temp" held the input lines.

Now, though, "current" holds the current input line, and "values" is an array of the registers to output.